# docs/internal_docs/internal_swagger.json
{
    "swagger": "2.0",
    "info": {
        "title": "Student data API",
        "contact": {},
        "version": "2.0"
    },
    "basePath": "/api/v1",
    "paths": {
...
        "/integrations/students/profile": {
            "get": {
                "security": [
                    {
                        "ApiKeyAuth": []
                    }
                ],
                "description": "Get student information by their email",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "External Students"
                ],
                "summary": "Get student by Email",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Student Email",
                        "name": "email",
                        "in": "query",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/integrations.StudentSuccessResponse"
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "$ref": "#/definitions/response.CommonErrorResponse"
                        }
                    },
                    "403": {
                        "description": "Forbidden",
                        "schema": {
                            "$ref": "#/definitions/response.CommonErrorResponse"
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "$ref": "#/definitions/response.CommonErrorResponse"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/response.CommonErrorResponse"
                        }
                    }
                }
            }
        },
        "/integrations/students/profiles": {
            "get": {
                "security": [
                    {
                        "ApiKeyAuth": []
                    }
                ],
                "description": "Get a paginated list of students with optional filters",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "External Students"
                ],
                "summary": "List students",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Search term for student name or number",
                        "name": "search",
                        "in": "query"
                    },
                    {
                        "type": "integer",
                        "description": "Filter by program ID",
                        "name": "program_id",
                        "in": "query"
                    },
                    {
                        "type": "integer",
                        "description": "Filter by gender ID",
                        "name": "gender_id",
                        "in": "query"
                    },
                    {
                        "type": "integer",
                        "description": "Filter by year level",
                        "name": "year_level",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "Order by field",
                        "name": "order_by",
                        "in": "query"
                    },
                    {
                        "type": "integer",
                        "description": "Page number for pagination",
                        "name": "page",
                        "in": "query"
                    },
                    {
                        "type": "integer",
                        "description": "Number of items per page for pagination",
                        "name": "page_size",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/integrations.ListStudentsSuccessResponse"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/response.CommonErrorResponse"
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "$ref": "#/definitions/response.CommonErrorResponse"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/response.CommonErrorResponse"
                        }
                    }
                }
            }
        },
        "/integrations/students/{studentNumber}": {
            "get": {
                "security": [
                    {
                        "ApiKeyAuth": []
                    }
                ],
                "description": "Get student information by student number",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "External Students"
                ],
                "summary": "Get student by student number",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Student number of the student",
                        "name": "studentNumber",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/integrations.StudentSuccessResponse"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/response.CommonErrorResponse"
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "$ref": "#/definitions/response.CommonErrorResponse"
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "$ref": "#/definitions/response.CommonErrorResponse"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/response.CommonErrorResponse"
                        }
                    }
                }
            }
        },
        "/integrations/students/{studentNumber}/addresses": {
            "get": {
                "security": [
                    {
                        "ApiKeyAuth": []
                    }
                ],
                "description": "Get all addresses of a student by their student number",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "External Students"
                ],
                "summary": "Get student addresses by student number",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Student number of the student",
                        "name": "studentNumber",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/integrations.StudentAddressSuccessResponse"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/response.CommonErrorResponse"
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "$ref": "#/definitions/response.CommonErrorResponse"
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "$ref": "#/definitions/response.CommonErrorResponse"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/response.CommonErrorResponse"
                        }
                    }
                }
            }
        },
        "/integrations/students/{studentNumber}/personal-info": {
            "get": {
                "security": [
                    {
                        "ApiKeyAuth": []
                    }
                ],
                "description": "Get personal information of a student by their student number",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "External Students"
                ],
                "summary": "Get personal information by student number",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Student number of the student",
                        "name": "studentNumber",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/integrations.StudentPersonalInfoSuccessResponse"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/response.CommonErrorResponse"
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "$ref": "#/definitions/response.CommonErrorResponse"
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "$ref": "#/definitions/response.CommonErrorResponse"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/response.CommonErrorResponse"
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
...
        "integrations.ListStudentsSuccessResponse": {
            "type": "object",
            "properties": {
                "data": {
                    "$ref": "#/definitions/integrations.OGOSListStudentsResponse"
                },
                "status": {
                    "allOf": [
                        {
                            "$ref": "#/definitions/response.JSendStatus"
                        }
                    ],
                    "example": "success"
                }
            }
        },
        "integrations.OGOSListStudentsResponse": {
            "type": "object",
            "properties": {
                "meta": {
                    "$ref": "#/definitions/structs.PaginationMetadata"
                },
                "students": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/integrations.OGOSStudentDTO"
                    }
                }
            }
        },
        "integrations.OGOSStudentAddressDTO": {
            "type": "object",
            "properties": {
                "addressType": {
                    "type": "string"
                },
                "barangay": {
                    "$ref": "#/definitions/locations.Barangay"
                },
                "city": {
                    "$ref": "#/definitions/locations.City"
                },
                "province": {
                    "$ref": "#/definitions/locations.ProvinceDTO"
                },
                "region": {
                    "$ref": "#/definitions/locations.Region"
                },
                "streetDetail": {
                    "$ref": "#/definitions/structs.NullableString"
                },
                "studentNumber": {
                    "type": "string"
                }
            }
        },
        "integrations.OGOSStudentDTO": {
            "type": "object",
            "properties": {
                "email": {
                    "type": "string"
                },
                "firstName": {
                    "type": "string"
                },
                "lastName": {
                    "type": "string"
                },
                "middleName": {
                    "$ref": "#/definitions/structs.NullableString"
                },
                "mobileNumber": {
                    "type": "string"
                },
                "program": {
                    "$ref": "#/definitions/students.Program"
                },
                "section": {
                    "type": "string"
                },
                "studentNumber": {
                    "type": "string"
                },
                "suffixName": {
                    "type": "string"
                },
                "yearLevel": {
                    "type": "integer"
                }
            }
        },
        "integrations.OGOSStudentPersonalInfoDTO": {
            "type": "object",
            "properties": {
                "dateOfBirth": {
                    "type": "string"
                },
                "gender": {
                    "$ref": "#/definitions/students.Gender"
                },
                "heightM": {
                    "type": "number"
                },
                "placeOfBirth": {
                    "type": "string"
                },
                "studentNumber": {
                    "type": "string"
                },
                "weightKg": {
                    "type": "number"
                }
            }
        },
        "integrations.StudentAddressSuccessResponse": {
            "type": "object",
            "properties": {
                "data": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/integrations.OGOSStudentAddressDTO"
                    }
                },
                "status": {
                    "allOf": [
                        {
                            "$ref": "#/definitions/response.JSendStatus"
                        }
                    ],
                    "example": "success"
                }
            }
        },
        "integrations.StudentPersonalInfoSuccessResponse": {
            "type": "object",
            "properties": {
                "data": {
                    "$ref": "#/definitions/integrations.OGOSStudentPersonalInfoDTO"
                },
                "status": {
                    "allOf": [
                        {
                            "$ref": "#/definitions/response.JSendStatus"
                        }
                    ],
                    "example": "success"
                }
            }
        },
        "integrations.StudentSuccessResponse": {
            "type": "object",
            "properties": {
                "data": {
                    "$ref": "#/definitions/integrations.OGOSStudentDTO"
                },
                "status": {
                    "type": "string",
                    "example": "success"
                }
            }
        },
        "locations.Barangay": {
            "type": "object",
            "properties": {
                "cityCode": {
                    "type": "integer"
                },
                "code": {
                    "type": "string"
                },
                "id": {
                    "type": "integer"
                },
                "name": {
                    "type": "string"
                }
            }
        },
        "locations.City": {
            "type": "object",
            "properties": {
                "code": {
                    "type": "string"
                },
                "district": {
                    "$ref": "#/definitions/structs.NullableString"
                },
                "id": {
                    "type": "integer"
                },
                "name": {
                    "type": "string"
                },
                "provinceCode": {
                    "$ref": "#/definitions/structs.NullableString"
                },
                "regionCode": {
                    "$ref": "#/definitions/structs.NullableString"
                },
                "type": {
                    "$ref": "#/definitions/structs.NullableString"
                },
                "zipCode": {
                    "$ref": "#/definitions/structs.NullableString"
                }
            }
        },
        "locations.ProvinceDTO": {
            "type": "object",
            "properties": {
                "code": {
                    "$ref": "#/definitions/structs.NullableString"
                },
                "id": {
                    "type": "integer"
                },
                "name": {
                    "$ref": "#/definitions/structs.NullableString"
                },
                "regionCode": {
                    "$ref": "#/definitions/structs.NullableString"
                }
            }
        },
        "locations.Region": {
            "type": "object",
            "properties": {
                "code": {
                    "type": "string"
                },
                "id": {
                    "type": "integer"
                },
                "name": {
                    "type": "string"
                }
            }
        },
        "response.CommonErrorResponse": {
            "type": "object",
            "properties": {
                "code": {
                    "type": "integer",
                    "example": 400
                },
                "data": {},
                "message": {
                    "type": "string",
                    "example": "Detailed error message"
                },
                "status": {
                    "type": "string",
                    "example": "fail"
                }
            }
        },
        "response.JSendStatus": {
            "type": "string",
            "enum": [
                "success",
                "fail",
                "error"
            ],
            "x-enum-varnames": [
                "StatusSuccess",
                "StatusFail",
                "StatusError"
            ]
        },
...
        "structs.NullableString": {
            "type": "object",
            "properties": {
                "string": {
                    "type": "string"
                },
                "valid": {
                    "description": "Valid is true if String is not NULL",
                    "type": "boolean"
                }
            }
        },
        "structs.PaginationMetadata": {
            "type": "object",
            "properties": {
                "page": {
                    "type": "integer"
                },
                "pagesSize": {
                    "type": "integer"
                },
                "total": {
                    "type": "integer"
                },
                "totalPages": {
                    "type": "integer"
                }
            }
        },
        "students.Gender": {
            "type": "object",
            "properties": {
                "id": {
                    "type": "integer"
                },
                "name": {
                    "type": "string"
                }
            }
        },
...
        "students.Program": {
            "type": "object",
            "properties": {
                "code": {
                    "type": "string"
                },
                "id": {
                    "type": "integer"
                },
                "name": {
                    "type": "string"
                }
            }
        },
...
    },
    "securityDefinitions": {
        "ApiKeyAuth": {
            "type": "apiKey",
            "name": "Authorization",
            "in": "header"
        }
    }
}